+++ CHANGELOG.md
@@ -13,6 +13,7 @@ You can also check the
 
 - Ability to download the Sunshine data as CSV
 - Improve map chart hovering with unit displays, text length adjustments and style changes
+- Operational standards charts show trends over time and peer comparison instead of histograms.
 
 # 2.40.4 - 2026-04-07